Refactor and reduce tests to show critical differences & set up for usage with export
